[PP-2639] Add a new report to the inventory and holds reports zip tha… (#2549)
…t contains a list of holds with no remaining licenses.

## Description
This PR adds a a report to the inventory and holds report zip that is
currently being generated. The report contains holds that are associated
with license pools without any owned copies and therefore will never be
convertible to loans.

## Motivation and Context
We need to do the backend work to generate a report for libraries that
lists all titles with holds but no remaining licenses. This report can
be modelled after the current holds report.
https://ebce-lyrasis.atlassian.net/browse/PP-2639
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
I updated the unit tests which now cover the new functionality. I will
validate on minotaur once it is approved and merged.
## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.

---------

Co-authored-by: Tim DiLauro <tdilauro@users.noreply.github.com>