Fix flaky TestCreateMarkUnavailableChord::test_exception test (#2555)
## Description

Update the test to try to make it not occasionally fail.

## Motivation and Context

I've seen this fail on several test runs lately like this:

https://github.com/ThePalaceProject/circulation/actions/runs/15881997329/job/44784547536

## How Has This Been Tested?

Both the old test and the new one run successfully for me locally. So
we'll probably have to just merge this one and see if we continue to see
this test fail occasionally or not.

- Running tests locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.