Deprecate Axis 360 Datasource (PP-1911) (#2553)
## Description

Update how deprecated data sources are handled:
- If a datasource using the old name exists, use it
- If it doesn't, create one using the new name

This will let us deprecate a datasource by:
1. Update `DEPRECATED_NAMES` to include the deprecated name
2. Put this into a release and roll it out everywhere
3. In a follow up, once we are running a version of the code the case
deal with both the new and old names everywhere, do a DB update that
updates the old name to the new name.
4. Remove the deprecated name from `DEPRECATED_NAMES`.

This also updates the user facing references to the datasource name to
display the new name, rather than the deprecated name.

This PR implements step 1 above. I will make a follow up PR to this one
that implements step 3 & 4, which can go in once this PR is in a
release.

## Motivation and Context

See: PP-1911. We want to display `Boundless` instead of `Axis 360` in
the apps.

## How Has This Been Tested?

- New unit tests
- Running old unit tests
- Light local testing
- Grepping for places that we are directly creating / loading
`DataSource` instead of using the `lookup` method.

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.