Update ability to deprecate identifiers (PP-1911) (#2560)
## Description

This PR updates how we deprecate `Identifier` objects to match how we
deprecate `DataSource` objects in
https://github.com/ThePalaceProject/circulation/pull/2553.

## Motivation and Context

Did this as part of the work on PP-1911. 

This PR puts everything in place, but DOES NOT deprecate `AXIS_360_ID`
because I'm concerned it will cause some issues in the apps with
identifiers changing. I'm going to bring this up in the tech strategy
meeting today.

Even if we don't immediately end up using this code, I think its still
worth merging in, since it:
- Aligns the deprecation strategy with `DataSource`
- Centralizes how we load identifiers passed into controllers

## How Has This Been Tested?

- Testing locally
- New unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.