[PP-2676] Eliminate deadlocks during presentation calculation by limi… (#2558)
…ting work updates to one per transaction.

This update includes a new test to verify that unprocessed
work_id/policy objects are put back in the redis set in the case of
failures.
<!--- Describe your changes -->

We're seeing deadlocks in the calculate_presentation task.
https://ebce-lyrasis.atlassian.net/browse/PP-2676
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

A new test was added to reflect changes.
We likely won't know if this update solves the problem until we deploy
to production.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [ ] All new and existing tests passed.