[PP-2689] stop using redis set for queuing presentation recalcs (#2567)
## Description
We were using a redis set to collect work id/presentation calculation
policy pairs in order to late process them in batches periodically. It
turned out that doing things in that way create a couple of problems:

1. Batching seemed to be causes a high rate of deadlocks.
2. We were not getting first in first out processing. Since we're using
a set, there is no guarantee that the oldest items were being processed
first. In cases where there was some backup of the queue this could
result in items not being processed for long periods of time.

Since we moved to using a single transaction for each work id /
presentation calc policy combo, using the set no longer made sense since
we can use the task queue itself for ordering calculation operations.

This PR does three things:

1.  It introduces a new task for processing a single work / policy pair.
2. It ensures that anything in the redis set is transferred to the
default queue for processing.
3. It ensures all new calculations are placed on the default queue.

In a follow up PR, we'll remove the redis set and deploy once all the
redis queues are empty.


## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2689
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
New unit tests.

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.