Include download link in Boundless DRM fulfillment (PP-2702) (#2576)
## Description

Include a link to the encrypted epub along with the boundless license
document when doing a boundless DRM fulfillment.

The new document looks like this:
```json
{
  "licenseDocument": {...},
  "links": [
    {
      "rel": "publication",
      "type": "application/epub+zip",
      "href": "http://blahblah"
    }
  ]
}
```

## Motivation and Context

Currently the apps don't have a way to get from the fulfillment document
to the encrypted epub, since the license document does not provide a
link. So we encapsulate the document, to provide the apps with all the
information they need to complete the fulfillment.

## How Has This Been Tested?

- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.