Make sure classifier lookup isn't dependent on import order (PP-2682) (#2575)
## Description

This PR creates a new function `lookup_classifier` that makes sure all
the classifiers are imported and registered when it is called.
Previously which classifiers were registered depended on import order,
which meant that changing imports could cause different subjects to be
classified.

## Motivation and Context

Looking into PP-2682 (and more directly SUP-366) I found that at some
point, our import order changed, and the `OverdriveClassifier` was not
getting added to the list of available classifiers.

This means that we were ignoring the most relevant Overdrive metadata,
which was causing bad classifications for Overdrive books.

## How Has This Been Tested?

- Running unit tests in CI
- Tested locally with some of the titles mentioned in SUP-366 and
verified that:
  - Before this change we were getting bad classifications
  - After this change we are getting much better classifications

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.