[PP-2705] remove coveragerecord adds and updates from calculate_work_presentation path. (#2581)
## Description
This PR removes the updates to the CoverageRecord table when an
edition's cover or metadata is recalculated.
<!--- Describe your changes -->

## Motivation and Context
Since we are using celery now to recalculate presentation information,
we no longer need to support Edition related coverage records.
https://ebce-lyrasis.atlassian.net/browse/PP-2705
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Obsolete unit tests have been removed. All current unit tests are
passing. I confirmed that everything appears to be working locally.
However I'll want to test on minotaur as well. Assuming all is well, I'd
like to create a hotfix release with this change and deploy it to
ga-georgia or vt-vermont before rolling it out production-wide.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.