Split up api/circulation.py (PP-2470) (#2584)
## Description

No code changes here, other then to:
- Split up `api/circulation.py` into `api/circulation/*`
- Remove unused `CirculationInfo` class
  - Also remove the test (the only place it was used)
- Rename `CirculationAPI` to `CirculationApiDispatcher` to make it more
clear what the distinction between it and implementations of
`BaseCirculationApi` are.

## Motivation and Context

In PP-2470 I am pulling apart the OPDS importer classes which implement
`BaseCirculationApi` and I thought having a `circulation` package would
be helpful, so I can move some of the reusable parts of those importers
into it.

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.