Remove OPDS validate scripts (PP-2470) (#2585)
## Description

Remove the `opds2_schema_validate` and `opds2_odl_schema_validate`
scripts. We have the same validation functionality available in
https://github.com/ThePalaceProject/palace-tools repo, so it doesn't
make sense to have it included here as well.

## Motivation and Context

In PP-2470 I'm reworking the OPDS importer classes, which these scripts
extend. Since the validation functionality exists elsewhere, I think it
makes more sense to remove it, then to update it to match the structure
of the new importers.

## How Has This Been Tested?

- Running unit tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.