Add disable_async_calculation flags in order to prevent infinite loop. (#2583)
## Description
This PR is a follow-on PR to correct a problem with
https://github.com/ThePalaceProject/circulation/pull/2581

In a nutshell, the `work.calculate_work_presentation ` celery task, when
recalculating the presentation edition for a pool with more than one
edition makes a call into BibliographicData.apply() which in turn
triggers a call to `work.calculate_work_presentation`, thus creating an
infinite loop. Before #2581, the loop would be broken because a check
was being made in the `apply()` method to see if the coverage record had
been recently updated. If not, it would skip it.

This PR adds a flag to tell the `apply` method not to trigger a
`work.calculate_work_presentation` celery task when it is called within
the context of a celery task.

The solution of passing a flag down the call stack is a little
inelegant, but it _is_ minimally invasive and minimizes the risk for
introducing new bugs.
<!--- Describe your changes -->

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2705
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I tested tested locally.
I added a new unit test to cover the flag.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.