Add IntegrationRegistry.configurations_query (PP-2470) (#2591)
## Description

Add two new functions:
- `IntegrationRegistry.configurations_query`
- This gives a sqlalchemy select statement to find all the
`IntegrationConfiguration` for a particular protocol in the registry.
- `Collection.select_by_protocol`
- Leverages `IntegrationRegistry.configurations_query` to provide a
select statement targeted at collections with a particular protocol.

And refactored some older code to use these new functions. Note: Not all
code has been updated, only code that was already doing the same thing.
Eventually it would be good to move all code using
`Collection.by_protocol` to `Collection.select_by_protocol`, but for now
I just added a comment saying that `select_by_protocol` should be used
for new code.

## Motivation and Context

In PP-2470 I need to query collections by protocol. We already do this
in the opds_odl celery task, and the boundless celery task, so I decided
to pull this out to a function, so its not being duplicated.

## How Has This Been Tested?

- New unit tests
- Running existing tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.