Update booleans in configuration forms to use bool values (#2602)
## Description

Update type hint, and update configuration form booleans to use boolean
values in options.

## Motivation and Context

In one of our config forms we were using "True" instead of "true" in
`options` which was causing the form not to work. Instead of mixing
string and boolean, just update them all use to use boolean values for
options to make it clear what we should be using.

## How Has This Been Tested?

- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.