Fix settings warning in logs (#2603)
## Description

Fix a warning that is occurring frequently in our logs:
> Configuration form item does not have a default value or factory and
yet its required property is set to False

## Motivation and Context

This warning was added in:
https://github.com/ThePalaceProject/circulation/pull/2239

But it is too broad, and is thrown for any required value that didn't
explicitly set its form item to required as well. We only want to have a
warning when we are explicitly setting a required item to not required.
To accomplish this we set the default to `SentinelType.NotGiven` so we
can distinguish between it not being set and being set to `False`
incorrectly.

## How Has This Been Tested?

- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.