Add some basic validation for boundless DRM parameters (#2614)
## Description

Add some validation for the parameters passed into boundless DRM

## Motivation and Context

See slack here:
https://lyrasis.slack.com/archives/CCY3PH8JD/p1752831605606329

After doing some debugging on parameters today, I thought it might be
helpful to do a bit of validation of the parameters before we pass them
on to boundless. Should make it easier to troubleshoot issues in the
future.

## How Has This Been Tested?

- New unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.