Allow circulation api to sort delivery mechanisms (PP-2744) (#2606)
## Description

Allow `BaseCirculationAPI` implementations to sort the list of delivery
mechanisms to be included in feeds.

This includes some refactoring to move the sorting that was being done
for `OPDS` integration's out to the appropriate API implementations,
instead of being tightly coupled to the `Collection`.

## Motivation and Context

In PP-2744 I need a setting to prioritize Boundless DRM. This provides
the hooks I need for that. That implementation will come in a future PR.

## How Has This Been Tested?

- New unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.