Fix issue with BibliographicData validator (PP-2722) (#2616)
## Description

Fix an issue where the `BibliographicData` published validator was
causing `BibliographicData` not to be able to be round tripped from
JSON.

To fix this I rolled back the changes from
https://github.com/ThePalaceProject/circulation/pull/2501 and took
another approach to solve the problem that PR was attempting to solve.

## Motivation and Context

In PP-2722, I need to be able to round trip `BibliographicData` to /
from JSON. When I tried to do this, I hit the issue I mentioned in this
comment:
https://github.com/ThePalaceProject/circulation/pull/2501#discussion_r2126673379

It looks like the `BibliographicData` published validator was added in
https://github.com/ThePalaceProject/circulation/pull/2501 to resolve
PP-2610.

Instead of solving this via a validator, I updated the OPDS importer
code to make sure what it passes into `BibliographicData` is the type
that is expected.

## How Has This Been Tested?

- Tested locally
- Once its deployed, I'll check the imports on minotaur and make sure
they are succeeding

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.