Avoid performing updates when no changes have occurred during present… (#2623)
…ation calculation to reduce probability of deadlocks.

## Description
On CMs with multiple overlapping Boundless collections we are seeing
long database lock acquisition times and Deadlocks when attempting to
update the editions table. This problem becomes more acute as the number
of overlapping collections increases.

This PR attempts address this problem by being smarter about not trying
to update the cover url and thumbnail information if it hasn't changed.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2725
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests are passing. There is no change in behavior expected at the
application level.

I'm not sure how to effectively test that non-data changing sql updates
are not being propagated to the database without having to dig into the
sqlalchemy layer. @jonathangreen or @tdilauro: if you think this level
of coverage is necessary, I'll spend some time looking into it.

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.