Revert "Bypass update when reaping boundless collections (#2601)" (#2624)
This reverts commit 9646a9a46ca5ccb54411198320c4c32a7c432ca2.

## Description
From my testing on Georgia, It appears that
https://github.com/ThePalaceProject/circulation/pull/2623 is a more
effective fix both for the problem of queue build up as well as that of
deadlocks in CMs with multiple overlapping Boundless collections.
Since this change (ie the one I am proposing to revert in this PR
[9646a9a46ca5ccb54411198320c4c32a7c432ca2]) opened the possibility of
missing updates, reverting it here makes sense.

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I deployed v32.0.2 + #2623 (which does not included the commit I'm
reverting here in this PR) to georgia and observed it for a couple of
hours and all looks good. While I did observe a deadlock, it was in a
different place in the code than the place we were seeing before.
Additionally the default queue, which previously would touch zero but
generally not stay there, is now staying at zero.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.