Trigger collection task from collection controller (PP-2722) (#2629)
## Description

Add a new method to trigger a collection indexing task when collection
settings are changed.

## Motivation and Context

I need this functionality for the PR I'm making for PP-2722. Breaking
this part out, because that PR is a beast, and I'd like to get some of
the pieces of it landed.

This doesn't actually do anything right now. I'm going to make a couple
PRs to implement the `import_task` method as a follow on to this.

## How Has This Been Tested?

- New tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.