Prevent unnecessary updates to target age. (#2626)
## Description
Following on the success of #2623, this PR aims to further reduce
database load by reducing the number of locks we're placing. We reduce
locks by not triggering sql updates when data does not change.

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
These are minor changes. We would expect the existing unit tests to
continue to pass. We'll need to run in the wild to see if we achieve the
desired reduction in database load.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.