Don't log warning when creating analytics event outside flask (#2641)
## Description

Don't log a warning when creating analytics events outside of a flask
context.

## Motivation and Context

While working on
https://github.com/ThePalaceProject/circulation/pull/2639 I had to dig
though a lot of our celery logs, and we are getting a large number of
warnings in the logs `Unable to resolve the user_agent`. This is
expected in a celery task, so I don't think it warrants a log message.

I updated the `RuntimeError` we would expect from flask to do nothing,
and updated the logging we do for any other exception to include a stack
trace with the message since it would be unexpected.

## How Has This Been Tested?

- New unit tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.