[PP-2704] Ignore contributors with blank display name values. (#2638)
## Description
Before this PR, opds2 imports were failing if a contributor's name was
blank. This condition can arise when metadata property that is mapped to
a contributor role, such as "translator", appears in the json with a
blank value.

As a result, after the contributors were parsed, we were seeing a
ValueError being thrown
[here](https://github.com/ThePalaceProject/circulation/blob/main/src/palace/manager/data_layer/contributor.py#L220)
which was causing the book import to fail (just for that book). But
since all of the books in the PressBoooks feed contained translator="",
all the books were failing.

This PR changes the behavior such that a WARNING is logged and the
nameless contributor ignored rather than recording a import failure for
the book in question. An exception would be justified if OPDS2 specified
that metadata properties MUST be non-blank. In fact the spec says that
value SHOULD NOT be blank.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2704
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit test added. 
Existing tests pass.
Tested manually in my local environment with with PressBooks OER
collection.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.