Convert OPDS2 importer to Celery Task (PP-2722) (#2635)
## Description

This pull request converts the OPDS2 importer functionality to run as a
Celery task.

Because the `OPD2` and the `OPDS2WithODL` importers intertwined, this PR
contains a lot of changes to the `OPDS2WithODL` importer as well. Even
though it was updated, the `OPDS2WithODL` importer wasn't converted to
Celery in this PR. The work here will make that easier when the time
comes, but I wanted to be able to roll out the importer updates one at a
time.

Conceptually I split what used to be the importer into two parts:
- `Importer` this provides the methods to do the actual import, used by
the Celery task
- `Extractor` this takes the OPDS and converts it into our internal data
model for later import

The extractor doesn't require any network requests or database access,
it purely does a mapping from one data model to the other. The importer
makes the network requests, and updates the database, or queues other
tasks to do so.

## Motivation and Context

Epic: PP-107 
Ticket: PP-2722

- Update OPDS2 importer to import with Celery task
- Update OPDS2WithODL importer so it doesn't depend on the OPDS2
importer
- This involved moving some of the methods from OPDS2 importer into
OPDS2WithODL importer.
- Since OPDS2WithODL is the next up for conversion, this situation won't
last too long.

## How Has This Been Tested?

- Local testing
- New unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.