Refactor our ServicesFixture to make sure a wired services container is always available (#2642)
## Description

This PR refactors our `ServicesContainer` fixture to make sure that
tests always have a wired services container available.

## Motivation and Context

Previously, if a test required a wired `services_container`, it had to
either wire it manually or use the `wired_services_container` fixture.
This approach was chosen because wiring the container is a relatively
expensive operation.

However, as we increasingly rely on the `services_container` and use it
for more dependency injection, managing it manually has become
cumbersome. We’re finding ourselves needing to request a wired container
in more and more places, which adds overhead and friction.

Specifically in
https://github.com/ThePalaceProject/circulation/pull/2643, because we
call up `Collection.data_source` in many places, that PR required adding
a lot of code to the tests, just to wire a services container in various
places.

This PR introduces an alternative approach:

* A session-level `services_container` is created and wired once.
* A function-level fixture then ensures the container is reset between
tests to avoid state leakage.

There is some risk of test data leaking if the container isn’t fully
reset, but the dependency injection framework provides helper methods to
assist with cleanup. If we encounter issues, we can enhance the session
fixture to handle additional reset logic.

Overall, I believe the convenience of always having a pre-wired
container outweighs the potential risk of state leakage between tests.

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.