Refactor how get a data_source for a collection (#2643)
## Description

Refactor `Collection.data_source` to delegate the lookup to
`BaseCirculationAPI.data_source`.

Note this is based on:
https://github.com/ThePalaceProject/circulation/pull/2642 and that PR
should go in first.

## Motivation and Context

This is the fix I would have liked to make instead of
https://github.com/ThePalaceProject/circulation/pull/2640 (PP-2758), but
because this builds on other unreleased code, this fix wouldn't be very
amiable to a hotfix release, so I'm making to PR as a follow up.

Takes a similar approach to
https://github.com/ThePalaceProject/circulation/pull/2606, where the
collection looks up the CirculationApi and delegate to it.

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.