Update the retry parameters for apply tasks (#2656)
## Description

Update the lock and retry parameters for the `apply` tasks.

## Motivation and Context

Last night after the roll out of v33.0.1, we had a few celery exceptions
like this:
```
Task apply.bibliographic_apply[7c5187ff-10de-40a3-b9ee-36e9da86e07f] raised unexpected: LockNotAcquired
```

There weren't too many of them, and I'd expect we did a huge amount of
importing, based on the timestamps being reset, but I think based on the
data in our logs, we can tune the retry parameters to make this less
likely to happen.

I set the `lock_timeout` to 5 minutes, since looking at the logs most of
these apply in less then a second, and the ones that take a long time
take 10 seconds. 5 minutes should be plenty of time for the timeout.

I also increased the number of retries, and the backoff, as well as
setting the maximum backoff. This should all give the tasks more time to
complete, and allow them to be better spaced out if there are competing
tasks.

## How Has This Been Tested?

- This is mostly configuration, so it shouldn't cause an issue. 
- Once this rolls out we'll keep monitoring exceptions to see if it
helps

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.